+++ lib/presentation/resources/widgets/molecules/location_widget.dart
@@ -95,84 +95,88 @@ class _LocationWidgetState extends State<LocationWidget> {
         const SizedBox(
           height: 12,
         ),
-        ThemeContainer(
-          height: 40,
-          child: Row(
-            children: [
-              ThemeText(
+        Row(
+          children: [
+            ThemeContainer(
+              height: 40,
+              customWidth: true,
+              width: 335,
+              child: ThemeText(
                 textString: _userLocation,
                 fontSize: 16,
                 fontWeight: FontWeight.w400,
                 textColor: TravalongColors.primary_text_bright,
+                height: 2, // todo: fjern, prøvede at få tekst i midten
               ),
-              const Spacer(), // moves button to right side
-              ElevatedButton(
-                // TODO: change ElevatedButton to custom button
-                onPressed: () async {
-                  // TODO: move logic to "Domain" layer
+            ),
+            //const Spacer(), // moves button to right side
+            const SizedBox(width: 2),
+            ElevatedButton(
+              // TODO: change ElevatedButton to custom button
+              onPressed: () async {
+                // TODO: move logic to "Domain" layer
 
-                  // check if location service is enabled on user device
-                  bool serviceEnabled =
-                      await Geolocator.isLocationServiceEnabled();
-                  if (!serviceEnabled) {
-                    return Future.error("Location services are disabled.");
-                  }
+                // check if location service is enabled on user device
+                bool serviceEnabled =
+                    await Geolocator.isLocationServiceEnabled();
+                if (!serviceEnabled) {
+                  return Future.error("Location services are disabled.");
+                }
 
-                  // check/get permission for location access
-                  LocationPermission permission =
-                      await Geolocator.checkPermission();
+                // check/get permission for location access
+                LocationPermission permission =
+                    await Geolocator.checkPermission();
 
+                if (permission == LocationPermission.denied) {
+                  permission = await Geolocator.requestPermission();
                   if (permission == LocationPermission.denied) {
-                    permission = await Geolocator.requestPermission();
-                    if (permission == LocationPermission.denied) {
-                      return Future.error("Location permissions are denied.");
-                    }
+                    return Future.error("Location permissions are denied.");
                   }
+                }
 
-                  if (permission == LocationPermission.deniedForever) {
-                    return Future.error(
-                        "Location permissions are permanently denied, cannot request permissions.");
-                  }
+                if (permission == LocationPermission.deniedForever) {
+                  return Future.error(
+                      "Location permissions are permanently denied, cannot request permissions.");
+                }
 
-                  // when permissions are granted we can
-                  // get the current position of device
-                  Position position = await Geolocator.getCurrentPosition(
-                      desiredAccuracy: LocationAccuracy.high);
+                // when permissions are granted we can
+                // get the current position of device
+                Position position = await Geolocator.getCurrentPosition(
+                    desiredAccuracy: LocationAccuracy.high);
 
-                  // get city and country based on coordinates
-                  List<Placemark> placemarks = await GeocodingPlatform.instance
-                      .placemarkFromCoordinates(
-                          position.latitude, position.longitude);
+                // get city and country based on coordinates
+                List<Placemark> placemarks = await GeocodingPlatform.instance
+                    .placemarkFromCoordinates(
+                        position.latitude, position.longitude);
 
-                  // Get the first placemark from the list
-                  Placemark placemark = placemarks[0];
+                // Get the first placemark from the list
+                Placemark placemark = placemarks[0];
 
-                  // Store the city and country in a string
-                  String? city = placemark.locality;
-                  String? country = placemark.country;
+                // Store the city and country in a string
+                String? city = placemark.locality;
+                String? country = placemark.country;
 
-                  setState(() {
-                    // update database with _userLocation
-                    fController.setDocFieldData(UserData.city, city);
-                    fController.setDocFieldData(UserData.country, country);
+                setState(() {
+                  // update database with _userLocation
+                  fController.setDocFieldData(UserData.city, city);
+                  fController.setDocFieldData(UserData.country, country);
 
-                    // refresh _userLocation display text
-                    fController.getDocFieldData(UserData.city).then((city) {
-                      fController
-                          .getDocFieldData(UserData.country)
-                          .then((country) {
-                        setState(() {
-                          _userLocation = "$city, $country";
-                          //debugPrint(_userLocation);
-                        });
+                  // refresh _userLocation display text
+                  fController.getDocFieldData(UserData.city).then((city) {
+                    fController
+                        .getDocFieldData(UserData.country)
+                        .then((country) {
+                      setState(() {
+                        _userLocation = "$city, $country";
+                        //debugPrint(_userLocation);
                       });
                     });
                   });
-                },
-                child: const Icon(Icons.update),
-              ),
-            ],
-          ),
+                });
+              },
+              child: const Icon(Icons.update),
+            ),
+          ],
         ),
       ],
     );
