+++ lib/presentation/screens/profile/my_profile_page.dart
@@ -1,5 +1,4 @@
 import "package:flutter/material.dart";
-import 'package:google_fonts/google_fonts.dart';
 import 'package:travalong/presentation/screens/profile/interests_subpage.dart';
 import 'package:travalong/presentation/resources/widgets/atoms/back_arrow.dart';
 import 'package:travalong/presentation/resources/widgets/atoms/safe_scaffold.dart';
@@ -8,8 +7,6 @@ import 'package:travalong/presentation/resources/widgets/molecules/interest_hobb
 import 'package:travalong/presentation/resources/widgets/molecules/location_widget.dart';
 import 'package:travalong/presentation/resources/widgets/molecules/sign_out_btn_widget.dart';
 import 'package:travalong/presentation/resources/widgets/molecules/theme_topbar.dart';
-
-import '../../resources/colors.dart';
 
 class MyProfilePage extends StatefulWidget {
   const MyProfilePage({super.key});
@@ -22,12 +19,10 @@ class _MyProfilePageState extends State<MyProfilePage> {
   @override
   Widget build(BuildContext context) {
     return SafeScaffoldNoNavbar(
-      topbar: ThemeTopBar(
+      topbar: const ThemeTopBar(
         title: "My Profile",
-        backArrow: const BackArrow(),
-        enableCustomButton: true,
-        customButtonWidget:
-            SaveButtonWidget(onPress: () => saveAction()), // todo: fix warning
+        backArrow: BackArrow(),
+        enableCustomButton: false,
       ),
       child: Padding(
         padding: const EdgeInsets.symmetric(horizontal: 14),
@@ -45,39 +40,6 @@ class _MyProfilePageState extends State<MyProfilePage> {
           ],
         ),
       ),
-    );
-  }
-
-  // * onPress parameter for SaveButtonWidget
-  void saveAction() {
-    // TODO: Save stuff to database
-  }
-}
-
-class SaveButtonWidget extends StatelessWidget {
-  final void onPress;
-
-  const SaveButtonWidget({
-    Key? key,
-    required this.onPress,
-  }) : super(key: key);
-
-  @override
-  Widget build(BuildContext context) {
-    return IconButton(
-      constraints: const BoxConstraints.expand(width: 80),
-      icon: Text(
-        'Save',
-        textAlign: TextAlign.center,
-        style: GoogleFonts.poppins(
-            fontSize: 20,
-            fontWeight: FontWeight.w500,
-            color: TravalongColors.secondary_10),
-      ),
-      onPressed: (() => {
-            onPress,
-            Navigator.pop(context) // return to previous page
-          }),
     );
   }
 }
